add validate token and update delete-polls

+++ src/services/vote.service.js
@@ -1,3 +1,4 @@
+import { query } from "express";
 import database from "../config/database.js";
 
 const CreatePolls = async( PollsBody ) => {
@@ -12,13 +13,12 @@ const CreatePolls = async( PollsBody ) => {
         console.log( " exist title", isExistPolls[0] );
         return { message: "EXIST POLLS" };
     }
-    const Polls = database.query(`INSERT INTO polls( userid ,title, isLock, createdAt ) VALUES( ?, ?, ?, ? )`, [ PollsBody.userid, PollsBody.title, PollsBody.isLock, PollsBody.createdAt ]);
-    if( Polls[0].length === 0 ) {
+    const Polls = await database.query(`INSERT INTO polls( userid ,title, isLock, createdAt ) VALUES( ?, ?, ?, ? )`, [ PollsBody.userid, PollsBody.title, PollsBody.isLock, PollsBody.createdAt ]);
+    if( Polls[0]?.affectedRows === 0 ) {
         console.log( Polls[0] );
         return { message : "CREATE POLLS FAIL", status: false  };
     } 
-    console.log( Polls[0] );
-    return { message: "CREATE POLLS SUCCESS", PollsID : Polls[0].insertId, status: true  };   
+    return { message: "CREATE POLLS SUCCESS", PollsID : Polls[0]?.insertId, status: true  };   
     }
     catch( error ) {
         console.log( error );
@@ -52,7 +52,7 @@ const LockPoll = async( PollID, UserID ) => {
         }
 
         const Polls = await database.query(`UPDATE polls SET isLock = 1 WHERE id = ?`, [ PollID ]);
-        if( Polls[0].length === 0 ) {
+        if( Polls[0]?.affectedRows === 0 ) {
             console.log( Polls[0] );
             return { message : "POLLS LOCK IS FAIL", status: false  };
         } 
@@ -91,7 +91,7 @@ const UnLockPoll = async( PollID, UserID ) => {
         }
 
         const Polls = await database.query(`UPDATE polls SET isLock = 0 WHERE id = ?`, [ PollID ]);
-        if( Polls[0].length === 0 ) {
+        if( Polls[0]?.affectedRows === 0 ) {
             console.log( Polls[0] );
             return { message : "POLLS LOCK IS FAIL", status: false  };
         } 
@@ -104,26 +104,51 @@ const UnLockPoll = async( PollID, UserID ) => {
     }
 }
 
-const DeletePoll = async( PollID ) => {
+const DeletePoll = async( PollID, UserID ) => {
     try {
         const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [PollID]);
         if( isExistPolls[0].length === 0 ) {
-            console.log( " not exist userid", isExistPolls );
+            console.log( " not exist pollID", isExistPolls );
             return { message: "NOT EXIST POLLSID"  };
         }
-        const _RemoveOption = await RemoveOption( PollID );
-        console.log( _RemoveOption )
-        if ( !_RemoveOption.status ) {
-            return { message : "DELETE USER OPTION FAIL", status: false  };
-        };
+
+        const isExistUser = await database.query(`SELECT * FROM users WHERE id = ?`, [UserID]);
+        if( isExistUser[0].length === 0 ) {
+            console.log( " not exist userid", isExistUser );
+            return { message: "NOT EXIST USERID" };
+        }
+
+        if( isExistPolls[0][0].userid !== UserID ) {
+            console.log( "NOT CREATOR", UserID, isExistPolls[0][0].userid )
+            return { message: "NOT CREATOR", status: false };
+        }
+
+        const isExistOption = await database.query(`SELECT * FROM options WHERE pollID = ?`, [ PollID ]);
+        if( isExistOption[0].length !== 0 ) {
+            const isExistUserOption = await database.query(`SELECT * FROM submition WHERE optionID = ?`, [ isExistOption[0][0].ID ]);
+            if( isExistUserOption[0].length !== 0 ) {
+                console.log( isExistUserOption )
+                const RemoveUserOption = await database.query(`DELETE FROM submition WHERE optionID = ?`, [ isExistOption[0][0].ID ]);
+                console.log( RemoveUserOption );
+                if( RemoveUserOption[0]?.affectedRows === 0 ) {
+                    console.log( "not have any vote", RemoveUserOption );
+                    return { message: "DELETE SUBMITION FAIL", status: false }
+                }
+            }
+            const RemoveOption = await database.query(`DELETE FROM options WHERE pollID = ?`, [ PollID ]);
+            if( RemoveOption[0]?.affectedRows === 0 ) {
+                console.log( "not have any option", RemoveOption );
+                return { message: "DELETE OPTION FAIL", status: false }
+            }
+        }
 
         const DeletePoll = await database.query(`DELETE FROM polls WHERE id = ?`, [ PollID ]);
-        if( DeletePoll[0].length === 0 ) {
+        if( DeletePoll[0]?.affectedRows === 0 ) {
             console.log( DeletePoll[0] );
             return { message : "DELETE POLLS FAIL", status: false  };
         }
 
-        console.log( DeletePoll[0] );
+        console.log( DeletePoll[0], RemoveOption[0] );
         return { message: "DELETE POLLS SUCCESS", status: true  };
     } catch (error) {
         console.log( error )
@@ -135,7 +160,7 @@ const AddOption = async( OptionBody ) => {
     try {
         const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [OptionBody.pollid]);
         if( isExistPolls[0].length === 0 ) {
-            console.log( " not exist userid", isExistPolls );
+            console.log( " not exist POLLSID", isExistPolls );
             return { message: "NOT EXIST POLLSID", status: false  };
         }
 
@@ -158,13 +183,14 @@ const AddOption = async( OptionBody ) => {
         }
 
         const Option = await database.query(`INSERT INTO options( pollID, title, createdAt ) VALUES(?,?,?)`, [ OptionBody.pollid, OptionBody.option, OptionBody.createdAt ]);
-        console.log( Option );
-        if( Option[0].length === 0 ) {
+        console.log( Option[0]?.affectedRows );
+        if( Option[0]?.affectedRows === 0 ) {
             console.log( Option );
             return { message : "ADD OPTION FAIL", status: false  };
         } 
-        console.log( isExistPolls[0], isPollsLock, isOptionTitleNull[0], isOptionExist[0], Option[0].length );
-        return { message: "ADD OPTION SUCCESS", OptionID: Option[0].insertId, status: true  };   
+        // console.log( isExistPolls[0], isPollsLock, isOptionTitleNull[0], isOptionExist[0], Option[0].length );
+        console.log( Option[0]?.insertId )
+        return { message: "ADD OPTION SUCCESS", OptionID: Option[0]?.insertId, status: true  };   
     } catch (error) {
         console.log( error )
         return { message: "ERROR ADD OPTION ", status: false  };
@@ -179,19 +205,30 @@ const RemoveOption = async( OptionID ) => {
             return { message: "NOT EXIST OPTION ID"}
         }
 
+        const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [ isExistOption[0][0].pollID ]);
+        if( isExistPolls[0].length === 0 ) {
+            console.log( " not exist POLLSID", isExistPolls );
+            return { message: "NOT EXIST POLLSID", status: false  };
+        }
+        const isPollsLock = isExistPolls[0][0].isLock
+        if( isPollsLock ) {
+            console.log( "POLLS IS LOCK", isPollsLock );
+            return { message: "POLLS IS LOCK" };
+        }
+
         const isExistUserOption = await database.query(`SELECT * FROM submition WHERE optionID = ?`, [OptionID]);
         if( isExistUserOption[0].length !== 0 ) {
             console.log( isExistUserOption )
             const RemoveUserOption = await database.query(`DELETE FROM submition WHERE optionID = ?`, [OptionID]);
             console.log( RemoveUserOption );
-            if( RemoveUserOption[0].length === 0 ) {
+            if( RemoveUserOption.affectedRows === 0 ) {
                 console.log( "not have any vote", RemoveUserOption );
                 return { message: "DELETE SUBMITION FAIL", status: false }
             }
         }
 
         const RemoveOption = await database.query(`DELETE FROM options WHERE ID = ?`, [OptionID]);
-        if( RemoveOption[0].length === 0 ) {
+        if( RemoveOption[0]?.affectedRows === 0 ) {
             console.log( "not have any option", RemoveOption );
             return { message: "DELETE OPTION FAIL", status: false }
         }
@@ -217,10 +254,21 @@ const Vote = async( VoteBody ) => {
             return { message: "NOT EXIST OPTION" };
         }
 
+        const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [ isExistOption[0][0].pollID ]);
+        if( isExistPolls[0].length === 0 ) {
+            console.log( " not exist POLLSID", isExistPolls );
+            return { message: "NOT EXIST POLLSID" };
+        }
+        const isPollsLock = isExistPolls[0][0].isLock
+        if( isPollsLock ) {
+            console.log( "POLLS IS LOCK", isPollsLock );
+            return { message: "POLLS IS LOCK" };
+        }
+
         const isExistVote = await database.query(`SELECT * FROM submition WHERE userID = ? AND optionID = ?`, [VoteBody.userid, VoteBody.optionid]);
         if( isExistVote[0].length === 0 ) {
             const vote = await database.query(`INSERT INTO submition( userID, optionID) VALUES(?,?)`, [ VoteBody.userid, VoteBody.optionid ]);
-            if( vote[0].length === 0 ) {
+            if( vote[0]?.affectedRows === 0 ) {
                 console.log( vote[0] );
                 return { message : "VOTE FAIL", status: false  };
             } 
@@ -247,10 +295,21 @@ const UnVote = async( UnVoteBody ) => {
             return { message: "NOT EXIST OPTION"};
         }
 
+        const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [ isExistOption[0][0].pollID ]);
+        if( isExistPolls[0].length === 0 ) {
+            console.log( " not exist POLLSID", isExistPolls );
+            return { message: "NOT EXIST POLLSID"};
+        }
+        const isPollsLock = isExistPolls[0][0].isLock
+        if( isPollsLock ) {
+            console.log( "POLLS IS LOCK", isPollsLock );
+            return { message: "POLLS IS LOCK"};
+        }
+
         const isExistVote = await database.query(`SELECT * FROM submition WHERE userID = ? AND optionID = ?`, [ UnVoteBody.userid, UnVoteBody.optionid]);
         if( isExistVote[0].length !== 0 ) {
             const unVote = await database.query(`DELETE FROM submition WHERE userID = ? AND optionID = ?`, [ UnVoteBody.userid, UnVoteBody.optionid ]);
-            if( unVote[0].length === 0 ) {
+            if( unVote[0]?.affectedRows === 0 ) {
                 console.log( unVote[0] );
                 return { message : "UNVOTE FAIL", status: false  };
             } 
@@ -267,7 +326,7 @@ const UnVote = async( UnVoteBody ) => {
 const GetUsersVoted = async( PollsID ) => {
     try {
         const isExistPolls = await database.query(`SELECT * FROM polls WHERE id = ?`, [PollsID]);
-        if( isExistPolls == 0 ) {
+        if( isExistPolls === 0 ) {
             console.log( " not exist userid", isExistPolls );
             return { message: "NOT EXIST POLLSID" };
         }
